RenderHealth move to correct directory - builds lib-tempo
